// src/lesson_7/level_3/interfaces.ts
// Level 2 - Интерфейсы для пропсов

export interface ButtonProps {
    text: string;
    color: string;
    // Callback - чтобы узнать в какой карточке было нажатие
    onClick?: () => void; // ? - необязательный пропс
}

export interface CardProps {
    id: number;
    price: number;
    title: string;
    text: string;
    image: string;
    alt: string;
    onAddToCart: (id: number, title: string, price: number) => void;
}

export interface CartItem {
    id: number;
    title: string;
    price: number;
    quantity: number;
    totalPrice: number;
}

export interface CartProps {
    items: CartItem[];
    total: number;
}